Move DataProvider into own class
